FragmentSearch
chuyển code từ SearchActivity sang FragmentSearch
