+++ app/src/main/java/com/example/prm392_musicapp/fragments/FragmentSearch.java
@@ -17,6 +17,7 @@ import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Toast;
 
 import com.example.prm392_musicapp.R;
 import com.example.prm392_musicapp.adapter.SearchAdapter;
@@ -45,6 +46,9 @@ public class FragmentSearch extends Fragment {
     private String mParam1;
     private String mParam2;
 
+    private  SearchAdapter searchAdapter;
+    private  SearchView  searchView;
+
     public FragmentSearch() {
         // Required empty public constructor
     }
@@ -67,7 +71,7 @@ public class FragmentSearch extends Fragment {
         return fragment;
     }
 
-  /*   @Override
+     @Override
    public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         if (getArguments() != null) {
@@ -81,16 +85,35 @@ public class FragmentSearch extends Fragment {
                              Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.search_page, container, false);
 
-          VideoDataUtils.searchVideoData("bac").observe(getViewLifecycleOwner(), new Observer<List<Item>>() {
-      @Override
-      public void onChanged(List<Item> items) {
-              //data ở trong cái items
-          Log.i("data",items.toString());
-      }
-  });
+//          VideoDataUtils.searchVideoData("bac").observe(getViewLifecycleOwner(), new Observer<List<Item>>() {
+//      @Override
+//      public void onChanged(List<Item> items) {
+//              //data ở trong cái items
+//          Log.i("data",items.toString());
+//      }
+//  });
 
-        SearchAdapter searchAdapter = new SearchAdapter(getListSong(),getActivity());
+       searchAdapter = new SearchAdapter(getListSong(),getActivity());
         RecyclerView revMusic = view.findViewById(R.id.rev_music);
+       searchView = view.findViewById(R.id.searchView);
+
+        searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
+            @Override
+            public boolean onQueryTextSubmit(String query) {
+                // Called when the user submits the search query
+                // You can perform your search operation here
+
+                return false;
+            }
+
+            @Override
+            public boolean onQueryTextChange(String newText) {
+                // Called when the user changes the query text
+                // You can perform real-time filtering or suggestions here
+                filterList(newText);
+                return true;
+            }
+        });
         RecyclerView.LayoutManager linearLayoutManager = new LinearLayoutManager(getActivity());
         revMusic.setLayoutManager(linearLayoutManager);
         revMusic.setAdapter(searchAdapter);
@@ -103,48 +126,49 @@ public class FragmentSearch extends Fragment {
 
     private List<Music> getListSong() {
         List<Music> list = new ArrayList<>();
-        list.add(new Music(R.drawable.nnca,  "Nơi này có anh", "Sơn Tùng"));
-        list.add(new Music(R.drawable.cadsv, "Chắc ai đó sẽ về", "Sơn Tùng"));
-        list.add(new Music(R.drawable.ctktvn, "Chúng ta không thuộc về nhau", "Sơn Tùng"));
-        list.add(new Music(R.drawable.atbe, "Âm thầm bên em", "Sơn Tùng"));
+        list.add(new Music(R.drawable.nnca,"Nơi này có anh","Sơn Tùng"));
+        list.add(new Music(R.drawable.cadsv,"Chắc ai đó sẽ về","Sơn Tùng"));
+        list.add(new Music(R.drawable.ctktvn,"Chúng ta không thuộc về nhau","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Âm thầm bên em","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Tình em là đại dương","Duy Mạnh"));
+        list.add(new Music(R.drawable.atbe,"Kiếp đỏ den","Duy Mạnh"));
+        list.add(new Music(R.drawable.atbe,"Phê","Duy Mạnh"));
+        list.add(new Music(R.drawable.atbe,"Lời xin lỗi của một dân chơi","Duy Mạnh"));
+        list.add(new Music(R.drawable.atbe,"Em của ngày hôm qua","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Chiếc khăn gió ấm","KHánh Phương"));
+        list.add(new Music(R.drawable.atbe,"Ta còn yêu nhau","Đức Phúc"));
+        list.add(new Music(R.drawable.atbe,"Không phải dạng vừa đâu","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Thái Bình mồ hôi rơi","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Lệ anh vẫn rơi","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Ấn nút nhớ thả giấc mơ","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Đừng về trễ","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Lạc trôi","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Run now","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Nắng âm xa dần","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Tình em là đại dương","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Cơn mưa ngang qua","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Mãi như ngày hôm qua","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Hãy trao cho anh","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Anh sai rồi","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Muộn rồi mà sao còn","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Có chắc yêu là đây","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"KHuôn mặt đáng thương","Sơn Tùng"));
+        list.add(new Music(R.drawable.atbe,"Một chút quên em thôi","Sơn Tùng"));
+
         return list;
     }
 
-    public boolean onCreateOptionsMenu(Menu menu) {
-
-        SearchAdapter searchAdapter = new SearchAdapter(getListSong(),getActivity());
-        appCompatActivity.getMenuInflater().inflate(R.menu.search_song_menu, menu);
-
-        SearchManager searchManager = (SearchManager) appCompatActivity.getSystemService(Context.SEARCH_SERVICE);
-        SearchView searchView = (SearchView) menu.findItem(R.id.action_search).getActionView();
-        searchView.setSearchableInfo(searchManager.getSearchableInfo(appCompatActivity.getComponentName()));
-        searchView.setMaxWidth(Integer.MAX_VALUE);
-
-        searchView.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
-            @Override
-            public boolean onQueryTextSubmit(String query) {
-                searchAdapter.getFilter().filter(query);
-                return false;
+    private  void filterList(String text){
+        List<Music> filteredList = new ArrayList<>();
+        List<Music> songList = getListSong();
+        for (Music music : songList){
+            if(music.getMusicName().toLowerCase().contains(text.toLowerCase())){
+                filteredList.add(music);
             }
-
-            @Override
-            public boolean onQueryTextChange(String newText) {
-                searchAdapter.getFilter().filter(newText);
-                return false;
-            }
-        });
-        return true;
+        }
+        if(!filteredList.isEmpty()){
+            searchAdapter.setFilteredList(filteredList);
+        }
     }
-
-//    SearchView searchView;
-
-//    public void onBackPressed() {
-//        ComponentActivity componentActivity = new ComponentActivity();
-//        if (!searchView.isIconified()) {
-//            searchView.setIconified(true);
-//            return;
-//        }
-//        super.onBackPressed();
-//    }*/
 
 }
