+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FFFFFF"
-    tools:context=".activities.HomeActivity">
+    tools:context=".activities.MainActivity">
 
     <FrameLayout
         android:id="@+id/fr_container"
