+++ app/src/main/res/layout/home_page.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FFFFFF"
-    tools:context=".activities.HActivity">
+    tools:context=".activities.MainActivity">
 
     <TextView
         android:id="@+id/textView3"
