+++ app/src/main/res/layout/library_page.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activities.LibraryActivity">
+    tools:context=".activities.MainActivity">
 
     <!-- TODO: Update blank fragment layout -->
