+++ app/src/main/res/layout/search_page.xml
@@ -7,20 +7,9 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FFFFFF"
-    tools:context=".activities.SearchActivity">
+    tools:context=".activities.MainActivity">
 
-    <!--    <TextView-->
-    <!--        android:id="@+id/textView4"-->
-    <!--        android:layout_width="match_parent"-->
-    <!--        android:layout_height="60dp"-->
-    <!--        android:gravity="center"-->
-    <!--        android:text="Search"-->
-    <!--        android:textAlignment="center"-->
-    <!--        android:textSize="20sp"-->
-    <!--        android:textStyle="bold"-->
-    <!--        app:layout_constraintEnd_toEndOf="parent"-->
-    <!--        app:layout_constraintStart_toStartOf="parent"-->
-    <!--        app:layout_constraintTop_toTopOf="parent" />-->
+
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rev_music"
@@ -49,45 +38,6 @@
         android:background="@drawable/btn_search_round"
         app:iconifiedByDefault="false"></androidx.appcompat.widget.SearchView>
 
-    <!--    <androidx.constraintlayout.widget.ConstraintLayout-->
-    <!--        android:layout_width="match_parent"-->
-    <!--        android:layout_height="wrap_content"-->
-    <!--        android:background="@drawable/btn_search_round"-->
-    <!--        app:layout_constraintEnd_toEndOf="parent"-->
-    <!--        app:layout_constraintStart_toStartOf="parent"-->
-    <!--        app:layout_constraintTop_toTopOf="parent">-->
-
-    <!--        <ImageView-->
-    <!--            android:id="@+id/imageView"-->
-    <!--            android:layout_width="50dp"-->
-    <!--            android:layout_height="30dp"-->
-    <!--            android:adjustViewBounds="true"-->
-    <!--            android:src="@drawable/baseline_search_24_dark"-->
-    <!--            app:layout_constraintBottom_toBottomOf="parent"-->
-    <!--            app:layout_constraintStart_toStartOf="parent"-->
-    <!--            app:layout_constraintTop_toTopOf="parent" />-->
-
-
-    <!--        <EditText-->
-    <!--            android:id="@+id/edt_search_text"-->
-    <!--            android:layout_width="0dp"-->
-    <!--            android:layout_height="wrap_content"-->
-    <!--            android:layout_weight="1"-->
-    <!--            android:background="@android:color/transparent"-->
-    <!--            android:imeOptions="actionSearch"-->
-    <!--            android:inputType="text"-->
-    <!--            android:lines="1"-->
-    <!--            android:hint="@string/action_search"-->
-    <!--            android:paddingLeft="8dp"-->
-    <!--            android:paddingTop="10dp"-->
-    <!--            android:paddingRight="8dp"-->
-    <!--            android:paddingBottom="10dp"-->
-    <!--            android:textSize="14sp"-->
-    <!--            app:layout_constraintEnd_toEndOf="parent"-->
-    <!--            app:layout_constraintStart_toEndOf="@+id/imageView"-->
-    <!--            app:layout_constraintTop_toTopOf="parent" />-->
-
-    <!--    </androidx.constraintlayout.widget.ConstraintLayout>-->
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
